CONSOLE-4837: Give up on fixing cycles
Note: the cycles produced by the frontend test are NOT as a result of new code introduced in this PR.

When we check for import cycles in the frontend test, we exclude any cycle with `get-active-plugins.js`. This is because there were a lot of cycles that were very difficult to fix due to the structure of how the console `PluginStore` is loaded.
Thus, any import cycle which includes the active plugins module doesn't count as a cycle in our eyes.

Meanwhile, each plugin entry point `plugin.tsx` is imported by the code produced by `get-active-plugins.js`. Now that we are trying to remove `plugin.tsx` from `console-app,` we are uncovering a bunch of import cycles that were previously not counted as `get-active-plugins.js` was previously in the cycle. Now that the cycles no longer include `get-active-plugins`, we are counting them.

Reverting only the changes to `plugin.tsx` shows this -- the detected cycles no longer appear.

Due to the amount of cycles that have to be dealt with (over 200), I want to address this in future work.

Claude prompt for future me:

This project currently has a bunch of import cycles. Import cycles can be detected by running "yarn check-cycles" in the `frontend` directly of this repository, and reading `frontend/.webpack-cycles`.

Your job is to fix all of the import cycles so that `yarn check-cycles` exits with code 0 instead of 1. You will fix cycles by rewriting imports--good examples can be observed in the last git commit, "CONSOLE-4837: Fix cycles". You will rewrite the imports by replacing their "index" or "barrel" file imports with the direct equivalent.

After rewriting imports, run `yarn check-cycles` again to confirm that all cycles have been fixed. Repeat until there are no more import cycles.